+++ app/commands/characters_context/dnd2024/update_command.rb
@@ -168,7 +168,9 @@ module CharactersContext
 
       def do_persist(input) # rubocop: disable Metrics/AbcSize
         input[:character].data =
-          input[:character].data.attributes.merge(input.except(:character, :avatar_file, :avatar_url, :name).stringify_keys)
+          input[:character].data.attributes.merge(
+            input.except(:character, :avatar_file, :avatar_url, :file, :name).stringify_keys
+          )
         input[:character].assign_attributes(input.slice(:name))
         input[:character].save!