+++ app/serializers/fate/character_serializer.rb
@@ -4,7 +4,10 @@ module Fate
   class CharacterSerializer < ApplicationSerializer
     include Deps[cache: 'cache.avatars']
 
-    attributes :provider, :id, :name, :created_at, :avatar
+    attributes :provider, :id, :name, :created_at, :avatar, :aspects, :phase_trio
+
+    delegate :data, to: :object
+    delegate :aspects, :phase_trio, to: :data
 
     def provider
       'fate'